OptionInsightMVP/requirements.txt
yfinance
pandas
numpy
scipy
requests
python-decouple
polygon-client
